+++ examples/resources/polaris_aws_account/resource.tf
@@ -113,3 +113,45 @@ resource "polaris_aws_account" "default" {
   }
 }
 
+# Enable Cloud Native Protection and Cyber Recovery Data Scanning with Outpost.
+resource "polaris_aws_account" "default" {
+  profile = "default"
+
+  cloud_native_protection {
+    permission_groups = [
+      "BASIC",
+    ]
+
+    regions = [
+      "us-east-2",
+      "us-west-2",
+    ]
+  }
+
+  cyber_recovery_data_scanning {
+    permission_groups = [
+      "BASIC",
+    ]
+
+    regions = [
+      "us-east-2",
+      "us-west-2",
+    ]
+  }
+
+  outpost {
+    outpost_account_id      = "123456789123"
+    outpost_account_profile = "outpost"
+
+    permission_groups = [
+      "BASIC",
+    ]
+  }
+}
+
+
+
+# The Couldformation stack ARN is available after creation
+output "stack_arn" {
+  value = polaris_aws_account.default.exocompute[0].stack_arn
+}